add templating menu
