+++ src/backoffice/templating/index.ts
@@ -1,4 +1,4 @@
-import { manifests as menuManifests } from './manifests';
+import { manifests as menuManifests } from './menu.manifests';
 import { manifests as templateManifests } from './templates/manifests';
 import { umbExtensionsRegistry } from '@umbraco-cms/extensions-api';
 import { ManifestTypes } from '@umbraco-cms/extensions-registry';
